CBFS: use cbfs_get_file_content whenever possible rather than cbfs_get_file
Number one reason to use cbfs_get_file was to get file length.
With previous patch no more need for this.

+++ src/vendorcode/google/chromeos/chromeos.c
@@ -82,7 +82,7 @@ int recovery_mode_enabled(void)
 }
 
 #if CONFIG_VBOOT_VERIFY_FIRMWARE
-void *vboot_get_payload(int *len)
+void *vboot_get_payload(size_t *len)
 {
 	struct vboot_handoff *vboot_handoff;
 	struct firmware_component *fwc;
